When shuffle is ENABLED. And you press PLAY in a dir, the selected song is now the first song to get played. To do this, we first remember which song that was selected, shuffle the list and then find the selected song in the shuffled list and play that.
This will have some interesting side effects if the song ends up for example
last in the shuffled list when we consider not always doing repeat-all as
then it'll reach the end of the playlist after one song.

We could optionally put the selected song first in the playlist and then
shuffle all the rest, as that would possibly be closer to what people will
assume it will do.


git-svn-id: svn://svn.rockbox.org/rockbox/trunk@1803 a1c6a512-1295-4272-9138-f99709370657