MEM should be passed in to this make, setting the size of the DRAM in MB.
We now preprocess the .lds file to allow macros/#ifdef etc and we now set
the DRAM size using a define. In combination with the recent configure fix,
this should allow us to build 8MB versions "out of the box".

The 'dist' target was removed.

'archos' was replaced with 'rockbox' in all the temporary file names when
building/linking. Plain cosmetic change.


git-svn-id: svn://svn.rockbox.org/rockbox/trunk@3345 a1c6a512-1295-4272-9138-f99709370657