Craig Sather's patch #749974:
Adds the ability to define 2 or more alternating lines on each line of the WPS
and the ability to control how long each alternating line is displayed before
switching to the next one.

Backward compatible with the current WPS specification language (existing WPS
files will display the same as they do now).

His HTML version of the docs is found at:

http://home.pacbell.net/csath/rockbox/details.html


git-svn-id: svn://svn.rockbox.org/rockbox/trunk@4229 a1c6a512-1295-4272-9138-f99709370657