Make the virtual keyboard easier to use for blind users. From FS#6324
When moving up/down and landing on the edit line, say "edit". Otherwise
it gets quite confusing.

When moving left and right on the edit line, sound a beep when the left
or right end is reached.

Before, it would say nothing either when the cursor is over a blank or
a char it doesn't know how to spell. So if we're on a blank spot, say
"blank".

Pass the whole utf8 char to be spelled.


git-svn-id: svn://svn.rockbox.org/rockbox/trunk@15060 a1c6a512-1295-4272-9138-f99709370657