Info screen: charging and battery.
For CHARGING_SIMPLE: if the charger is inserted, then indicate that the
battery is charging, and do not show the battery level (which should be
bogus anyway). This is the only place the charger status is reported for
blind users.

Allow reporting charge_state for CHARGING_MONITOR: the header file says
it could be CHARGING or TRICKLE. Hope this is right. Interestingly,
that's how the #ifdefs were for the voice part.

Have the voice match the displayed info.  In particular do not report
battery level when in a charging state.
Use the new talk time unit for the battery remaining duration.
If battery_level() is 0 and no info is available, then say "blank"
instead of nothing.
LANG_BATTERY_CHARGE: more verbose voice entry.


git-svn-id: svn://svn.rockbox.org/rockbox/trunk@15419 a1c6a512-1295-4272-9138-f99709370657