More voice for the radio: accept FS#6331.
In the radio screen, speak the mode (preset or scan) when switching mode.
When switching stations with the radio paused/muted, speak the current
preset (number, frequency if preset is default, or spell, depending
on talk_file setting). Also says the frequency if in scan mode.

Change the default preset created by autoscan to be an empty string; change
the preset menu callback to substitute the frequency when the preset name
is empty.

Add a voice callback for the preset menu.


git-svn-id: svn://svn.rockbox.org/rockbox/trunk@15507 a1c6a512-1295-4272-9138-f99709370657