Implement %Ss for playback speed
This is similar to %Sp for pitch.  The conditional form %?Ss<...> is also
supported in the same way as FS#10680; you can use

    %?Ss<%xdAa|%xdAb|>

to show an icon of a tortoise (subpicture 1 of image A) or a hare
(subpicture 2 of image A) when not playing at the normal speed.

Flyspray: FS#10681
Author: Junio C Hamano



git-svn-id: svn://svn.rockbox.org/rockbox/trunk@23591 a1c6a512-1295-4272-9138-f99709370657