Change several platform names to match configure names.
Rbutil's (internal) platform names are now identical to the configure
model names except for the cases where further differentiation is
needed. In those cases the platform name is now the configure name with
an additional postfix, separated by a dot. This makes it possible
getting the configure name from the platform name.

git-svn-id: svn://svn.rockbox.org/rockbox/trunk@24371 a1c6a512-1295-4272-9138-f99709370657