Brickmania: Revert to using intersection checks for testing if powerup hit paddle
r24925 introduced a potential bug: the collision detection with the paddle is
not using the line intersection checks and consequently could cause the powerup
to be missed if the drop speed is fast enough (which varies depending on the
target because of the speed scaling).

Thanks for Karl Kurbjun for noticing and alerting!


git-svn-id: svn://svn.rockbox.org/rockbox/trunk@24934 a1c6a512-1295-4272-9138-f99709370657