The zip target isn't a good prerequisite for another make target so remove it.
zip has no prerequisites, so it's always made, which leads to targets depending on zip to be also always made.
In the end that means that libmisc.so is always unzipped even if it didn't change.

OTOH it means that make zip is now needed explicitly; so it now goes like 'make && make zip && make apk

git-svn-id: svn://svn.rockbox.org/rockbox/trunk@28068 a1c6a512-1295-4272-9138-f99709370657