Backport r29548: autoresume: Manual: Clarify precedence rules for resume positions
Resume positions stored in bookmarks take precedence over autoresume.

Technically, the resume position for the "Resume Playback" function
(which is also saved across power-off/power-on) also takes precedence
over autoresume.  However, there is no way for a user to tell as it is
identical to the autoresume position for the last-played track, so
don't highlight it in the manual.

git-svn-id: svn://svn.rockbox.org/rockbox/branches/v3_8@29549 a1c6a512-1295-4272-9138-f99709370657