Add script to spellcheck the manual LaTeX sources.
This script wraps around aspell and calls it for the file passed on the command
line or runs aspell on all TeX files found in the manual folder (except
preamble.tex). aspell is called interactively. Note: there is currently no way
to interrupt checking if running on the complete folder.

git-svn-id: svn://svn.rockbox.org/rockbox/trunk@30013 a1c6a512-1295-4272-9138-f99709370657