Fix FS#12239, data abort after going from FMS to WPS.
The data abort was in find_handle() because current_handle was corrupted. This
most probably happened due to the FMS allocating Radio AA on the buffer without
resetting the buffer, corrupting buffering/playback state.

The fix is to grab buffer control explicitely.

r30308 isn't actually related, regardless of the task title.

git-svn-id: svn://svn.rockbox.org/rockbox/branches/v3_9@30343 a1c6a512-1295-4272-9138-f99709370657