Add a gitattributes file for the migration.
When we start using git, we will be doing automatic line ending conversion
in a nice uniform way: everything is expected to have LFs, rather than the
current odd patchwork of svn:eol-style attributes.

This .gitattributes file configures git to normalise all line endings to LF
on checkin in files it believes are text (and its heuristics appear to be
very good). It also explicitly declares a number of file types to be binary
(which both disables line ending conversion for them and also prevents git
attempting to display a textual diff of them).

I have already normalised all files in the repository now to have the correct
line endings, so this should not do anything bad for anyone who is already
using git.



git-svn-id: svn://svn.rockbox.org/rockbox/trunk@31104 a1c6a512-1295-4272-9138-f99709370657