Fix up tools/version.sh for git transition.
Since we will be transitioning to git, stop considering svn the primary source.
If HEAD is a git-svn revision, return the svn revision number as before
(this means that if you check out old versions they keep their previously
canonical version numbers) but if it's not, then just print the short SHA1.


git-svn-id: svn://svn.rockbox.org/rockbox/trunk@31480 a1c6a512-1295-4272-9138-f99709370657