Remove duplicate note from manual.
The note on how to recursively insert files into the dynamic playlist is
present twice so remove one of them. Since this makes the "Useful Hints"
subsection empty remove that as well, there are no hints left that aren't
merged into the text. (FS#12504 by Jean-Louis Biasini).

While at it remove some hard coded newlines. We really don't want them, if we
want a different spacing for new paragraphs this needs to get adjusted in the
preamble, not after each paragraph.

git-svn-id: svn://svn.rockbox.org/rockbox/trunk@31517 a1c6a512-1295-4272-9138-f99709370657