Fix building mkamsboot on OS X.
mkamsboot requires libucl to be linked. Since the introduction of libtools.make
the OS X specific universal library isn't triggered anymore so the libucl built
uses the wrong architecture. Rockbox Utility builds libucl by itself and still
triggers the universal library rule.

Since ar can create fat archives but not operate on them adjust the ar call to
not try to update the archive -- the archive is created anyway, so asking for
an update is not really necessary. Remove any old archive first to make sure
we're not trying to update one, since that would now fail. As a result the OS X
specific hackery to build both ppc and i386 in a single call isn't necessary
anymore.

git-svn-id: svn://svn.rockbox.org/rockbox/trunk@31585 a1c6a512-1295-4272-9138-f99709370657