Add alternate firmware filenames for e200v2.
Depending on the firmware version the filename is different, so it's necessary
to look for all the possible filenames when searching the file inside the zip.
The player happily accepts any of the filenames, so (as before) the first one
is used. Additional firmware filenames might be necessary for other players as
well.

Change-Id: If78444a8d9b7fe167ce0be1d58407038a4f9052b