Set HOME_DIR to "/". Should fix FS#12500.
The reasons for not using / don't seem to be valid (It can lead to //,
true, but not on any system where // is actually special, and
create_*_filename() handles / fine albeit not perfectly (more //)).

Handling /./ is desirable, but we can't afford to leave all automatic
filename generation broken indefinitely while people discuss possible
approaches.

Change-Id: I6718ea28d7c91e19f7da89f76c8cefd92e12fe5e