Add preliminary support for loopback-mounting a disk image.
This isn't very useful as such, but with some more work this could be
used to export arbitrary disk images as USB devices, possibly even (with
some more USB work) CD-ROM devices. With this sort of feature, it would
become possible to carry random bootable rescue images on a DAP.

What works now is mounting an image within rockbox on a non-hotswap
target. The image is also exported over USB currently, but this is
dangerous since the main disk (which has the image) is also exported.
Also, this really only works by accident because unmounting a filesystem
doesn't forcibly close file handles.

What needs more work:
- HAVE_HOTSWAP currently that only one driver supports hotswappable
  disks on a given target. This needs to be cleaned up.
- the USB code needs a mechanism to only export selected disks. This
  might actually be useful for other usecases as well (e.g. it might
  be possible to export the SD card slot while playing audio from the
  internal storage), but that would need even more work

Change-Id: Id4c7bc416641db70eb1ab7ff2ff8dffe03860dd2