Remove ata_removable() and ata_present() from multidriver code.
ata_removable() and ata_present() no longer exist (ATA drives are never
removable with our current drivers), so storage.c shouldn't try to call
them from the wrapper functions. This was never noticed because these
wrapper functions are only used for multidriver code, which happens not
to be used with ATA currently.

Change-Id: Icb5e8cb27cdbef3edc0e51c35dc40dadf4f9de29