Force settings touchscreen mode in menus.
Plugins revert to grid mode currently. If they create a list the list ends
up in grid mode as well even if they use point mode due to settings in the core.
The expectation is that the lists use whatever mode they also use in the core.

Force setting in menu to have them behave the same in core and plugins.

Change-Id: I9534b2f21ccfa920eb2e48f66b04ecbda7a59011