Update patch set 2
Patch Set 2:

This is the wrong fix. the code already calls lcd_getstringsize() so instead of basing the L/ife part on the lcd width, base it on the font size (i.e not a compile time option). when the game starts check the width of the string "Life: xx" and if its over some threshold drop it to L:

Patch-set: 2
Reviewer: Gerrit User 1000011 <1000011@583131e1-1a33-4f2b-b8ec-82434880f207>
Label: Tested=0