Update patch set 1
Patch Set 1:

This doesn't work on Android, unless I made an error in enabling it the build. When making a test panicf it just segfaulted.

I didn't test further. It *should* work if the unwinder doesn't use instructions unavialable in user mode.

As the patch as-is doesn't compile the unwinder for !PLATFORM_NATIVE I didn't check "fails on target".

Patch-set: 1