Python module to get (some) files from the repository.
With git it's not possible to simply export a set of paths from the server
directly. This module offers a similar functionality, but requires a local
clone of the repository. It calls git directly for maximum portability.

This is a prerequisite to making the deployment script used for Rockbox Utility
work again.

Change-Id: I1aca8ddd40d16b6268c1d6a2fc23ac59e964c3c6