Update patch set 1
Patch Set 1:

I agree, and I think that this could assist in removing a few of those options. "Insert After..." could replace "Insert Next" and "Insert Last", also it's possible to replace "Insert Shuffled" with "Insert Shuffled After...".
What I'm trying to is that "Insert After..." and "Insert Before..." are more generic than the current options, but I guess it would be a bit less convenient. 
But we could minimize the menu to:
- Insert
- Insert Shuffled After...
- Insert After...
- Queue
- Queue Shuffled After...
- Queue After...
- Play Next

All this without losing any functionality at all, maybe a little bit of convenience that was in "Insert Next" and "Insert Last" but overall I think that the flexibility is worth it. Plus maybe we don't have to delete the current menu options but instead make it possible to choose what menu items you'd like in the settings.

Patch-set: 1
Reviewer: Gerrit User 1000089 <1000089@583131e1-1a33-4f2b-b8ec-82434880f207>
Label: Code-Review=0
Label: Tested=0